delegate player state handlers to encapsulating context
the context should be modifiable globally through the app, this is done to reduce repeated querying of the same data. all state has been moved from the component to the context to allow the rest of the app to mutate the state as required.

the helper functions (handlePlayPause, useEffect) will be moved to the player context.